requirements.txt
appnope==0.1.3
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.1
Django==4.1.6
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.5
executing==1.2.0
idna==3.4
inflection==0.5.1
ipython==8.10.0
itypes==1.2.0
jedi==0.18.2
Jinja2==3.1.2
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
numpy==1.24.2
openapi-codec==1.3.2
packaging==23.0
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.36
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
simplejson==3.18.3
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
traitlets==5.9.0
uritemplate==4.1.1
urllib3==1.26.14
wcwidth==0.2.6
